requirements.txt
aiodns==3.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiomultiprocess==0.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
argcomplete==3.6.2
asteval==1.0.5
attrs==25.3.0
bc-detect-secrets==1.5.44
bc-jsonpath-ng==1.6.1
bc-python-hcl2==0.4.2
beartype==0.21.0
beautifulsoup4==4.13.4
boolean.py==5.0
boto3==1.35.49
botocore==1.35.99
cached-property==2.0.1
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
checkov==3.2.440
click==8.2.1
click-option-group==0.5.7
cloudsplaining==0.7.0
colorama==0.4.6
ConfigArgParse==1.7.1
contextlib2==21.6.0
contourpy==1.3.2
cycler==0.12.1
cyclonedx-python-lib==7.6.2
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
docker==7.1.0
dockerfile-parse==2.0.1
dpath==2.1.3
fonttools==4.58.2
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==7.2.1
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
junit-xml==1.9
kiwisolver==1.4.8
lark==1.2.2
license-expression==30.4.1
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.3
multidict==6.4.4
networkx==2.6.3
numpy==2.3.0
openai==1.86.0
orjson==3.10.18
packageurl-python==0.13.4
packaging==23.2
pillow==11.2.1
ply==3.11
policy_sentry==0.13.2
prettytable==3.16.0
propcache==0.3.2
py-serializable==1.1.2
pycares==4.8.0
pycep-parser==0.5.1
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
rdflib==7.1.4
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rpds-py==0.25.1
rustworkx==0.16.0
s3transfer==0.10.4
schema==0.7.5
semantic-version==2.10.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
spdx-tools==0.8.3
tabulate==0.9.0
termcolor==2.3.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
unidiff==0.7.5
uritools==5.0.0
urllib3==1.26.20
wcwidth==0.2.13
xmltodict==0.14.2
yarl==1.20.1
zipp==3.23.0
fpdf
pandas 
